+++ src/main/java/nekonic/models/Stock.java
@@ -4,12 +4,12 @@ import java.sql.Timestamp;
 
 public class Stock {
     private final String ticker;
-    private final double currentPrice;
+    private final int currentPrice;
     private final String historicalPrices;
     private final int volume;
     private final Timestamp updatedAt;
 
-    public Stock(String ticker, double currentPrice, String historicalPrices, int volume, Timestamp updatedAt) {
+    public Stock(String ticker, int currentPrice, String historicalPrices, int volume, Timestamp updatedAt) {
         this.ticker = ticker;
         this.currentPrice = currentPrice;
         this.historicalPrices = historicalPrices;
@@ -21,7 +21,7 @@ public class Stock {
         return ticker;
     }
 
-    public double getCurrentPrice() {
+    public int getCurrentPrice() {
         return currentPrice;
     }
